ml_model/requirements.txt
face-recognition
face-recognition-models
pillow
termcolor
numpy
